refactor(adr-018): extract runThreeSessionTddFromCtx to orchestrator-ctx.ts
