+++ src/tdd/index.ts
@@ -12,7 +12,8 @@ export {
   verifyTestWriterIsolation,
   verifyImplementerIsolation,
 } from "./isolation";
-export { runThreeSessionTdd, runThreeSessionTddFromCtx } from "./orchestrator";
+export { runThreeSessionTdd } from "./orchestrator";
+export { runThreeSessionTddFromCtx } from "./orchestrator-ctx";
 export { cleanupProcessTree, getPgid } from "./cleanup";
 export type { VerifierVerdict, VerdictCategorization } from "./verdict";
 export { VERDICT_FILE, readVerdict, cleanupVerdict, categorizeVerdict } from "./verdict";
